backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai==1.12.0

# Vector database
chromadb==0.4.22

# Data processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
markdown==3.5.2
pyyaml==6.0.1

# Image processing
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
aiofiles==23.2.1
tenacity==8.2.3
